// src/enums/water/waterTypes.ts
export enum WATER_TYPES {
  CUSTOM_WATER = 'CUSTOM_WATER',
  THIRD_WAVE_WATER_CLASSIC_LIGHT_ROAST_PROFILE = 'THIRD_WAVE_WATER_CLASSIC_LIGHT_ROAST_PROFILE',
  THIRD_WAVE_WATER_MEDIUM_ROAST_PROFILE = 'THIRD_WAVE_WATER_MEDIUM_ROAST_PROFILE',
  THIRD_WAVE_WATER_DARK_ROAST_PROFILE = 'THIRD_WAVE_WATER_DARK_ROAST_PROFILE',
  THIRD_WAVE_WATER_ESPRESSO_MACHINE_PROFILE = 'THIRD_WAVE_WATER_ESPRESSO_MACHINE_PROFILE',
  THIRD_WAVE_WATER_COLD_BREW_PROFILE = 'THIRD_WAVE_WATER_COLD_BREW_PROFILE',
  THIRD_WAVE_WATER_LOW_ACID_PROFILE = 'THIRD_WAVE_WATER_LOW_ACID_PROFILE',
  PURE_COFFEE_WATER = 'PURE_COFFEE_WATER',
}